Extract "install"s...
